+++ src/besselj.jl
@@ -24,7 +24,7 @@
 #
 
 """
-    besselj0(x::T) where T <: Union{Float32, Float64}
+    besselj0(x::T) where T <: Union{Float32, Float64, ComplexF32, ComplexF64}
 
 Bessel function of the first kind of order zero, ``J_0(x)``.
 
@@ -176,6 +176,14 @@ function _besselj1(x::Float32)
         return p * s
     end
 end
+
+#####
+#####  Implementation for complex arguments
+#####
+
+besselj0(z::T) where T <: Union{ComplexF32, ComplexF64} = besseli0(z/im)
+besselj1(z::T) where T <: Union{ComplexF32, ComplexF64} = besseli1(z/im) * im 
+
 
 #                  Bessel functions of the first kind of order nu
 #                               besselj(nu, x)